Move hooks from Conservation efforts component to container
