+++ src/components/landscape-sidebar/conservation-efforts-widget/conservation-efforts-widget-component.jsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from 'react';
-import { loadModules } from '@esri/react-arcgis';
+import React from 'react';
 import PieChart from 'components/pie-chart';
 import CheckboxGroup from 'components/checkbox-group';
 import { 
@@ -7,8 +6,6 @@ import {
   PROTECTED
 } from 'components/landscape-sidebar/conservation-efforts-widget/conservation-efforts-widget-selectors';
 import styles from './conservation-efforts-widget-styles.module.scss';
-
-const findInDOM = (id) => document.getElementById(id);
 
 const ConservationEffortsDescription = ({ allProp, rawData }) => {
   return (
@@ -20,8 +17,6 @@ const ConservationEffortsDescription = ({ allProp, rawData }) => {
 };
 
 const ConservationEffortsWidget = ({
-  setConservationEfforts,
-  terrestrialCellData,
   dataFormatted,
   colors,
   rawData,
@@ -31,53 +26,6 @@ const ConservationEffortsWidget = ({
   activeSlices,
   toggleLayer
 }) => {
-  const [conservationPropsLayer, setConservationPropsLayer] = useState(null);
-
-  useEffect(() => {
-    if (!conservationPropsLayer) {
-      loadModules(["esri/layers/FeatureLayer"]).then(([FeatureLayer]) => {
-        const consPropLayer = new FeatureLayer({
-          url: "https://services9.arcgis.com/IkktFdUAcY3WrH25/arcgis/rest/services/ConsProp/FeatureServer"
-        });
-        setConservationPropsLayer(consPropLayer)
-      });
-    }
-  }, []);
-
-  const queryParams = conservationPropsLayer && conservationPropsLayer.createQuery();
-
-  const orangeActive = alreadyChecked['Protected areas'];
-  const yellowActive = alreadyChecked['Community areas'];
-
-  useEffect(() => {
-    const svg = findInDOM('conservation-widget');
-    const orangeSlice = findInDOM(colors[PROTECTED]);
-    const yellowSlice = findInDOM(colors[COMMUNITY_BASED]);
-
-    if (svg && orangeSlice) {
-      if (orangeActive && yellowActive && orangeSlice && yellowSlice) {
-        // bring both to front
-        svg.appendChild(yellowSlice);
-        svg.appendChild(orangeSlice);
-      } else if (yellowActive && yellowSlice && !orangeActive) {
-        svg.appendChild(yellowSlice);
-      } else if (!yellowActive && orangeSlice && orangeActive) {
-        svg.appendChild(orangeSlice);
-      } else {
-        svg.appendChild(orangeSlice);
-      }
-    }
-  }, [orangeActive, yellowActive])
-
-  useEffect(() => {
-    if (terrestrialCellData && queryParams) {
-      queryParams.where = `CELL_ID IN (${terrestrialCellData.map(i => i.CELL_ID).join(', ')})`;
-      conservationPropsLayer.queryFeatures(queryParams).then(function(results){
-        const { features } = results;
-        setConservationEfforts(features.map(c => c.attributes));
-      });
-    }
-  }, [terrestrialCellData])
 
   return (
     <>
